requirements_gnn.txt
# --- Core numeric stack (base for many deps) ---
numpy==1.26.4
scipy==1.13.1
pandas==2.2.3
matplotlib==3.9.4
scikit-learn==1.6.1

# --- Jupyter ecosystem (will pull most jupyter_* packages) ---
jupyter==1.1.1
jupyterlab==4.4.0
ipywidgets==8.1.6

# --- Graph / GNNs (depends on TF/Keras, numpy, scipy) ---
spektral==1.3.1

# --- Engineering / FEM / geometry ---
calfem-python==3.6.11

plotly==5.19.0
